requirements.txt
torch>=1.9.0
numpy>=1.21.0,<2.0.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
pyyaml>=5.4.0

